pyproject.toml
[project]
name = "image-ftp-sender"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "cffi>=1.17.1",
    "colorama>=0.4.6",
    "cryptography>=44.0.2",
    "loguru>=0.7.3",
    "paramiko>=3.5.1",
    "pillow>=11.1.0",
    "pycparser>=2.22",
    "pyexiftool>=0.5.6",
    "pynacl>=1.5.0",
    "tqdm>=4.67.1",
]
